Update FindGLEW to create an imported target
As it does not ship with a cmake config, we cannot use that. Instead
updated the FindGLEW to mimic that behavior and export a GLEW::glew
target.

On linux, it does ship with a pkgconfig file. On that platform it just
uses cmake's PkgConfig module to find the target and then alias it to
GLEW::glew.