Add namespace libwethair to Core files and update App to reflect that
Wrapped all the files in Core with a libwethair namespace. Then
updated all files in App to use that namespace. Either explicitly in
headers or a just `using namespace libwethair` in source files.

I avoided using namespace libwethair in source files in Core, as at
least gcc strips the namespace of the symbols it creates when using
that. Which is causing linking issues.