Make Threads::Threads public for WetHairCore in the CMakeLists.txt
Otherwise it will cause a compilation error when the WetHairCore is
built as a shared library. I missed this the first time as when
building that statically, cmake will propagate _all_ dependencies to
the dependents. Given that a static library does not have concept of
NEEDED as a shared one does.