Add cmake/Findstb.cmake
Needed if finding the stb dependency instead of downloading it.

Reason for using `STB_IMAGE_WRITE_ROOT` and not `STB_ROOT` for finding
`stb_image_write.h`, is that some linux distributions (e.g. GNU Guix)
have split the stb into multiple packages. So this is future proofing
it somewhat in case libWetHair needs to use other headers from stb.
Given that each header needs to be searched for using find_path, to
find the correct include dir as they can be in different directories.