Add cmake option LIBWETHAIR_FIND_DEPENDENCIES
Which allows to find dependencies instead of downloading them. This is
off by default.

Downloading and vendoring dependencies does not fit well with package
managers on linux. The dependency graph is hidden from the view of the
package manager. Making it both wasteful (both disk space and compile
time) and more opaque when it comes to debugging dependency related
issues.

This option will allow the best of both worlds. If you are building
this without a package manager, you can leave it off. But if you are
packaging this up for a package manager, you can enable this and leave
the dependencies for the package manager.