Write to a log if pressure solve failed
Not ideal to spew all this to standard out, especially when the system
is large.  Better to write this to a log file if possible.  That way
you do not need to worry about blowing passed any scroll back limit or
bringing slower terminals to a crawl.

It will fallback to writing to standard error if it fails to open
the log file.