Momentum explosion detection (#20)
* Add max_velocity_ratio to WetHairParameter

Which will be used to detect when simulations goes unstable and
explodes, which can cause massive memory consumption and bring down
the machine.

* Add max velocity ratio check in momentum update in CylindricalShallowFlow

This can produces velocity spikes for certain scenarios when the
substep for the reduced liquid is too low.  It has the chain
reaction that hair strands explodes getting stretch 1000x which
libWetHair will generate dripping particles from.  That in turn
generates massive amount of fluid particles, which can easily exceed
100GB in ram usage from one frame to the next.